修改新增会员员工提成逻辑

+++ app/src/main/java/com/wycd/yushangpu/bean/EmplMsg.java
@@ -24,7 +24,7 @@ public class EmplMsg {
     private int EM_TipFastConsume;//	快速消费提成
     private int EM_TipTimesConsume;//	签到提成
     private int EM_TipComboConsume;//	套餐消费提成
-    private int staffProportion;//提成比例或者固定提成金额
+    private double staffProportion;//提成比例或者固定提成金额
     private boolean ischose;
 
     public boolean isIschose() {
@@ -211,11 +211,11 @@ public class EmplMsg {
         this.EM_TipComboConsume = EM_TipComboConsume;
     }
 
-    public int getStaffProportion() {
+    public double getStaffProportion() {
         return staffProportion;
     }
 
-    public void setStaffProportion(int staffProportion) {
+    public void setStaffProportion(double staffProportion) {
         this.staffProportion = staffProportion;
     }
 }
